+++ src/compas_rcf/utils/util_funcs.py
@@ -163,12 +163,12 @@ def ensure_frame(frame_like):
     )
 
 
-def get_offset_frame(origin_frame, distance):
+def get_offset_frame(frame, distance):
     """Offset a frame in its Z axis direction.
 
     Parameters
     ----------
-    origin_frame : `class`:compas.geometry.Frame
+    frame : `class`:compas.geometry.Frame
         Frame to offset
     distance : float
         Translation distance in mm
@@ -177,7 +177,7 @@ def get_offset_frame(origin_frame, distance):
     -------
     `class`:compas.geometry.Frame
     """
-    offset_vector = origin_frame.zaxis * distance * -1
+    offset_vector = frame.zaxis * distance * -1
     T = cg.Translation(offset_vector)
 
-    return origin_frame.transformed(T)
+    return frame.transformed(T)